feat(#809): move serve composition root to cmd/vibewarden/
Move all adapter-wiring code from internal/app/serve/ into
cmd/vibewarden/wiring_serve*.go so that internal/app/ no longer
imports any adapter packages. Delete the internal/app/serve/
package entirely.

Update the CI coverage step to drop the now-unnecessary
grep -v exclusion for internal/app/serve$.
