+++ .github/workflows/ci.yml
@@ -81,12 +81,10 @@ jobs:
           cache: true
 
       - name: Run coverage
-        # internal/app/serve is the composition root (wires adapters directly).
-        # Tracked for relocation to cmd/ in #809; excluded from unit coverage
-        # until that refactor lands, since wiring code is inherently exercised
-        # by integration tests and binary start-up, not unit tests.
+        # Composition-root wiring now lives in cmd/vibewarden/ (moved in #809),
+        # which is outside the coverage scope. No exclusion filter needed.
         run: |
-          PKGS=$(go list ./internal/domain/... ./internal/app/... | grep -v 'internal/app/serve$')
+          PKGS=$(go list ./internal/domain/... ./internal/app/...)
           go test -coverprofile=coverage.out -covermode=atomic $PKGS
 
       - name: Check coverage threshold
