+++ cmd/vibewarden/serve_test.go
@@ -3,18 +3,16 @@ package main
 import (
 	"context"
 	"testing"
-
-	appserve "github.com/vibewarden/vibewarden/internal/app/serve"
 )
 
 func TestRunServe_MissingConfig(t *testing.T) {
-	// RunServe should return an error when given a path to a non-existent config
+	// runServe should return an error when given a path to a non-existent config
 	// file that is not a standard search path (explicit path forces a load attempt).
-	err := appserve.RunServe(context.Background(), appserve.Options{
-		ConfigPath: "/nonexistent/path/to/vibewarden.yaml",
-		Version:    "test",
+	err := runServe(context.Background(), serveOptions{
+		configPath: "/nonexistent/path/to/vibewarden.yaml",
+		version:    "test",
 	})
 	if err == nil {
-		t.Error("RunServe() expected error for missing explicit config file, got nil")
+		t.Error("runServe() expected error for missing explicit config file, got nil")
 	}
 }
