+++ cmd/vibewarden/wiring_serve_helpers_test.go
@@ -1,4 +1,4 @@
-package serve
+package main
 
 import (
 	"testing"
