+++ cmd/vibewarden/wiring_serve_logger_test.go
@@ -1,4 +1,4 @@
-package serve
+package main
 
 import (
 	"context"
